+++ poc/classes/Axiom.py
@@ -1,13 +1,20 @@
 from poc.classes.AuxInterpretation import AuxInterpretation
+from poc.classes.AuxRuleDependencies import AuxRuleDependencies
 
 
 class Axiom(AuxInterpretation):
 
     def __init__(self, parse_list: list, parsing_info: AuxInterpretation):
-        self.clone(parsing_info)
+        super().__init__(parsing_info.get_ast_info(), parsing_info.get_errors())
+        self.predicate = None
         self.aggregate_previous_rules(parse_list,
-                                      ["axiom", "ax", "postulate", "post", "AxiomHeader"],
+                                      AuxRuleDependencies.dep["Axiom"],
                                       self.rule_aggregator)
 
     def rule_aggregator(self, rule: str, parsing_info: AuxInterpretation):
-        pass
+        if rule == "AxiomHeader":
+            self.stop_aggregation = True
+        elif rule == "Signature":
+            pass
+        elif rule == "AxiomBlock":
+            self.predicate = parsing_info.predicate