+++ poc/classes/RightBound.py
@@ -11,7 +11,7 @@ from poc.classes.AuxRuleDependencies import AuxRuleDependencies
 class RightBound(AuxInterpretation):
 
     def __init__(self, parse_list: list, parsing_info: AuxInterpretation):
-        self.clone(parsing_info)
+        super().__init__(parsing_info.get_ast_info(), parsing_info.get_errors())
         self.info = parsing_info
         self.bound_included = True
         self.aggregate_previous_rules(parse_list,
@@ -20,6 +20,6 @@ class RightBound(AuxInterpretation):
     def rule_aggregator(self, rule: str, parsing_info: AuxInterpretation):
         if rule == "ExclamationMark":
             pass
-        if rule == "RightBracket":
+        elif rule == "RightBracket":
             self.bound_included = False