+++ poc/classes/Xid.py
@@ -5,7 +5,7 @@ from poc.classes.AuxRuleDependencies import AuxRuleDependencies
 class Xid(AuxInterpretation):
 
     def __init__(self, parse_list: list, parsing_info: AuxInterpretation):
-        self.clone(parsing_info)
+        super().__init__(parsing_info.get_ast_info(), parsing_info.get_errors())
         self.aggregate_previous_rules(parse_list, AuxRuleDependencies.dep["XId"], self.rule_aggregator)
 
     def rule_aggregator(self, rule: str, parsing_info: AuxInterpretation):